requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==2.3.7
email-validator==2.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.22
WTForms==3.0.1
Jinja2==3.1.2
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.1
pandas==2.1.1
matplotlib==3.8.0
numpy==1.26.0
bcrypt==4.0.1
gunicorn==21.2.0
wordcloud
seaborn
dask[complete]==2023.12.1
distributed==2023.12.1
neo4j==5.14.1
py2neo==2021.2.4